% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{create_folder}
\alias{create_folder}
\title{Create a new folder inside the documents storage directory.}
\usage{
create_folder(name)
}
\arguments{
\item{name}{folder name}
}
\value{
TRUE if folder is created
}
\description{
Create a new folder inside the documents storage directory.
}
