% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{document}
\alias{document}
\title{Get a single document by its unique AnythingLLM document name}
\usage{
document(name)
}
\arguments{
\item{name}{Unique document name to find (name in /documents)}
}
\value{
A list for document
}
\description{
Get a single document by its unique AnythingLLM document name
}
