% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{document_remove}
\alias{document_remove}
\title{Permanently remove documents from the system.}
\usage{
document_remove(names)
}
\arguments{
\item{names}{Array of document names to be removed permanently.}
}
\value{
TRUE if success
}
\description{
Folder name is required to remove a document
}
