% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{documents}
\alias{documents}
\title{List of all locally-stored documents in instance}
\usage{
documents()
}
\value{
A list of all documents
}
\description{
List of all locally-stored documents in instance
}
