% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{documents_raw_text}
\alias{documents_raw_text}
\title{Upload a file by specifying its raw text content and metadata values without having to upload a file.}
\usage{
documents_raw_text(title, text, ...)
}
\arguments{
\item{title}{file title}

\item{text}{raw text of file}

\item{...}{other metadata with key and values}
}
\value{
The document metadata
}
\description{
Upload a file by specifying its raw text content and metadata values without having to upload a file.
}
