% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{embeddings}
\alias{embeddings}
\title{Add or remove documents from a workspace by its unique slug.}
\usage{
embeddings(slug, adds = NULL, deletes = NULL)
}
\arguments{
\item{slug}{Unique slug of workspace to find}

\item{adds}{files to add}

\item{deletes}{files to delete}
}
\value{
TRUE if embeddings updated
}
\description{
Add or remove documents from a workspace by its unique slug.
}
