% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{move_files}
\alias{move_files}
\title{Move files within the documents storage directory.}
\usage{
move_files(files, folder)
}
\arguments{
\item{files}{files to move including the folder name}

\item{folder}{new folder name}
}
\value{
TRUE if files are moved into new folder
}
\description{
Move files within the documents storage directory.
}
