% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{request}
\alias{request}
\title{Perform a request to TiddlyWiki WebServer}
\usage{
request(method = "GET", path = "/", body = NULL, auto_unbox = TRUE)
}
\arguments{
\item{method}{The method in the httr package, e.g. GET, POST}

\item{path}{The path of request}

\item{body}{the query body}

\item{auto_unbox}{whether auto unbox for json body}
}
\value{
The contents of response
}
\description{
Perform a request to TiddlyWiki WebServer
}
