% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{workspace}
\alias{workspace}
\title{Get a workspace by its unique slug.}
\usage{
workspace(slug)
}
\arguments{
\item{slug}{workspace slug}
}
\value{
A list for workspace
}
\description{
Get a workspace by its unique slug.
}
