% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{workspace_delete}
\alias{workspace_delete}
\title{Deletes a workspace by its slug.}
\usage{
workspace_delete(slug)
}
\arguments{
\item{slug}{workspace slug}
}
\value{
OK if deleted
}
\description{
Deletes a workspace by its slug.
}
