% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{workspace_new}
\alias{workspace_new}
\title{Create a new workspace}
\usage{
workspace_new(name)
}
\arguments{
\item{name}{workspace name}
}
\value{
A list for new workspace
}
\description{
Create a new workspace
}
